2025/11/23/practical-zero-trust-patterns-aiot-edge-devices/requirements.txt
cryptography>=41.0.0
paho-mqtt>=1.6.1
pyyaml>=6.0
pytest>=7.4.0
pytest-cov>=4.1.0

